requirements.txt
streamlit
jinja2
instructor
google-generativeai